Split tests in multiple files
