+++ package.json
@@ -54,8 +54,8 @@
     "in-publish": "^2.0.0",
     "nyc": "^11.4.1",
     "standard-version": "^4.0.0",
-    "tap-spec": "^4.1.0",
-    "tape": "^4.0.0",
+    "tap-spec": "^4.1.1",
+    "tape": "^4.8.0",
     "webpack": "^1.13.1"
   }
 }
